Various minor changes.

+++ include/kernels/gemm.hh
@@ -60,7 +60,8 @@ class gemm {
   /** Calculate a checksum from the result matrix C. */
   constexpr double calcChecksum() {
     // Checksum for GEMM calculated by summing all four corners of C together
-    return (C_[0] + C_[m_ - 1] + C_[(m_ * (n_ - 1))] + C_[m_ * n_ - 1]);
+    return ((double)C_[0] + (double)C_[m_ - 1] + (double)C_[(m_ * (n_ - 1))] +
+            (double)C_[m_ * n_ - 1]);
   }
 
  protected:
@@ -70,12 +71,12 @@ class gemm {
     srand(SEED);
     for (int y = 0; y < m_; y++) {
       for (int x = 0; x < k_; x++) {
-        A_[y * k_ + x] = (T)((rand() % 100) / 7);
+        A_[y * k_ + x] = (T)((double)(rand() % 100) / 7.0);
       }
     }
     for (int y = 0; y < k_; y++) {
       for (int x = 0; x < n_; x++) {
-        B_[y * n_ + x] = (T)((rand() % 100) / 3);
+        B_[y * n_ + x] = (T)((double)(rand() % 100) / 3.0);
       }
     }
   }
